Uses external module for VPC
By using an external, battle tested, module for the VPC we ansure that most good patterns and standards for both terraform and AWS are in place. This also helps people with previous experience with the module to contribute.

This PR also allows to use existing VPC and subnet in the main example.